+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Unreleased
+
+## Enhancements
+
+* Adds new run creation attributes: `allow-empty-apply`, `terraform-version`, `plan-only` by @sebasslash [#482](https://github.com/hashicorp/go-tfe/pull/447)
+
 # v1.6.0
 
 ## Enhancements